Syncer refactoring (porting to event bus in progress)
